backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
azure-identity>=1.17.0
python-dotenv>=1.0.1
azure-search-documents>=11.6.0
azure-core>=1.30.0
requests>=2.32.0
numpy>=1.26.0
aiortc>=1.9.0
